Update CONTRIBUTING.md
The section for the commit message format is too generic,
so this commit adds more details and a clear example of
how a commit message should be structured.

This commit also makes it clear in CONTRIBUTING.md that
JIRA should be used instead of GH issues and encourages
contributors to seek feedback and discussion on the proposed
changes early.

patch by João Reis; reviewed by TBD for CASSGO-10